feat(): changes to email pop up
* feat(): email and api calls for send email

* feat():subject and textarea added in email modal

* feat():api call

* feat(): changes to email pop up

+++ src/components/Admin/CRMTab/ProductQuantityEditor.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 export const ProductQuantityEditor = React.forwardRef((props, ref) => {
-    const [value, setValue] = React.useState(parseInt(props.value));
+    const [value, setValue] = React.useState(parseFloat(props.value));
     const refInput = React.useRef(null);
 
     React.useEffect(() => {
